Updating key/manifest generation
In the project we will grab the keys.json and use the first key to generate the manifest and key.js file.

On the project side, we are no longer exporting the value in keys.js, instead opting for a single constant that get’s exposed on the background. This alleviates the need to include the key during bundling so we can simplify the build process for multiple targets.