BREAKING CHANGE: Update build process to allow for multiple targets (#5)
* Consolidating duplicate code

- Adjusting folder structure to share code
- Adjusting run to resolve to general script
- Adding project to process.env.PROJECT
- Adjust general scripts to pick up project name

* Adding shared utilities

* Updating paths for new api key build structure

* Removing report and check tooling

Because we are potentially building for multiple targets, these tools were not being employed in a manner that leveraged their utility.  Removing them for now until we can update and leverage them properly.

* Updating Build / Adding Deploy

Updating build function to handle the initial compilation process.  Part of this includes leveraging common utilities.

Also adding a deploy function that generates key files and manifests based on an array of keys in the project directory.  This allows us to target multiple, variable browsers without additional specific config at the build level.

* Removing check for keys.js file

We are adjusting this requirement so we can target multiple browsers and make the key file generation a bit more straight forward.

* Updating key/manifest generation

In the project we will grab the keys.json and use the first key to generate the manifest and key.js file.

On the project side, we are no longer exporting the value in keys.js, instead opting for a single constant that get’s exposed on the background. This alleviates the need to include the key during bundling so we can simplify the build process for multiple targets.

* Updating readme for more clarity

Adding a message at the top of setup explaining that this repo is not necessary as a direct download.

Also adjusted the formatting to be more in line with the rest of the code base (Line breaks at 100)

* Updating package version

This is a breaking change as it requires a new keys.json in the src directory of the project and removes the need for a keys.js file in /common

* Update Pocket New Tab settings for multi-build

* Update copyLocales to check for locales folder

* Updating Manifest to check for locales

If there are no locales we will assume the manifest description should not be modified.

* Changing default dev directory to be more clear

* Updating lock file

* Re-Adding clearing of console

* Updating formatting automatically with jsPrettier